pyproject.toml
[project]
name = "scraper-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "saqibmurtaza", email = "saqibmurtazakhan@gmai.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.1",
    "hatchling>=1.27.0",
    "literalai>=0.1.201",
    "oauth2client>=4.1.3",
    "openai-agents>=0.0.15",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "selenium>=4.32.0",
]

[project.scripts]
scraper-multi-agent = "scraper_multi_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
